Don’t auto-close in the console input
E.g. if the student types `{` and presses enter, that should be that.

Extracts the creation of an ACE editor with universal defaults into a
module.
