+++ src/components/Editor.jsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ACE from 'brace';
 import bindAll from 'lodash/bindAll';
 import get from 'lodash/get';
 import throttle from 'lodash/throttle';
 import noop from 'lodash/noop';
-import createAceSessionWithoutWorker
-  from '../util/createAceSessionWithoutWorker';
+import {createAceEditor, createAceSessionWithoutWorker} from '../util/ace';
 
 import 'brace/ext/searchbox';
 import 'brace/mode/html';
@@ -98,16 +96,10 @@ class Editor extends React.Component {
 
   _setupEditor(containerElement) {
     if (containerElement) {
-      this._editor = ACE.edit(containerElement);
-      this._editor.$blockScrolling = Infinity;
+      this._editor = createAceEditor(containerElement);
       this._startNewSession(this.props.source);
-      this._disableAutoClosing();
       this._resizeEditor();
       this._editor.on('focus', this._resizeEditor);
-      this._editor.setOptions({
-        fontFamily: 'Inconsolata',
-        fontSize: '14px',
-      });
     } else {
       this._editor.destroy();
     }
@@ -121,13 +113,8 @@ class Editor extends React.Component {
     }
   }
 
-  _disableAutoClosing() {
-    this._editor.setBehavioursEnabled(false);
-  }
-
   _startNewSession(source) {
     const session = createAceSessionWithoutWorker(this.props.language, source);
-    session.setUseWrapMode(true);
     session.on('change', () => {
       this.props.onInput(this._editor.getValue());
     });
