+++ src/css/application.css
@@ -368,6 +368,7 @@ body {
   display: flex;
   flex-direction: column;
   font-family: 'Inconsolata';
+  font-size: 14px;
 }
 
 .editors__editor-container {
@@ -474,6 +475,7 @@ body {
   display: flex;
   flex-direction: column-reverse;
   font-family: 'Inconsolata';
+  font-size: 14px;
   max-height: 100%;
   overflow: auto;
 }
